Restore using @media as default context when validating <mf-*>
It was removed in 7623480 (and worked by accident before) but is
required for Window.matchMedia(), Element.media, MediaList.*Medium(),
which cannot provide a context representing @media to the parser entry
point (with the media query list provided as input).

It could be fixed by accepting a top-level rule name in createContext()
but this would not work for Element.media, which should theoretically
have no access to createContext(). However using a style rule as the
default context can be inappropriate and this should be fixed later
somehow.