Fix parsing element-dependent numeric substitutions
This is not specified but mix, random, tree counting functions, should
not be accepted in contexts that are not associated to an element, such
as media queries, except those that do not need to resolve their value
(@supports). Other numeric substitutions should be valid in any context.

Relative values like in sign(1em), which has type <number> and can be
combined with any other numeric types, should probably be resolved based
on initial property values when there is no associated element
(w3c/csswg-drafts#10853).