Fix validating @property
Computationally dependent values are now searched in function and simple
block values.

It is not clear whether keywords that resolve to an absolute value based
on another property, in a specific context, like smaller in font-size,
should be considered as computationally dependent. Current version of
Chrome and FF do not, so they are no longer invalid with a non-universal
syntax, for now.

As per WPT, <*-percentage> is now considered computationally
independent, and registering a custom property with the universal syntax
and a CSS-wide keyword as the initial value now fails
(w3c/css-houdini-drafts#1076), which was already implemented for
@property but not CSS.registerProperty().

WPT also expects var()-containing value to be invalid with the universal
syntax, but the reason is unclear so it is not implemented, for now.