Carry the input string to lower parsing levels
Custom property declarations must be assigned the original text of their
value in order to be serialized exactly as specified by the author, ie.
with any consecutive whitespaces and comments.

This was already implemented for custom property declarations in style
rules but not in the prelude of @container and only partially for the
prelude of @supports.