Avoid re-parsing matched values for validation
The validation result could otherwise be tagged with CSS types twice.
For exemple, 1px matching <dimension> produced by <mf-value> is
validated against the feature name value definition, and tagged a second
time with <dimension>, this time produced by <length>.

This did not appear in the tests because (1) this only happened with a
single component value (2) media features were the only cases but the
validation functions return the matched value, which is incorrect tho
and will be fixed in the next commit.