Refactorize context creation
As hinted in 9c8a8e5, the default context is no longer a style rule, to
avoid accepting element-dependent values in inappropriate contexts.

The context can now be created with a list of rule names, whose first
entry is the name of the top-level rule wrapping the next rule, and so
on. The last entry is the current context rule name.

The context model might be revisited in future commits.