Validate rules and declarations while consuming them
This is a significant and long overdue commit that changes the parsing
model for rules and declarations, which are now validated in the
corresponding CSS Syntax algorithms, conforming to a corresponding
change in the specification (w3c/csswg-drafts#8834) a few months ago,
giving up processing all parts of a style sheet with the same parser.

This implementation directly returns a CSSOM representation for rules,
which means that their constructor parses its block contents from the
token stream. The CSSOM entry point "parse a CSS rule" is removed; it is
assumed that CSSOM needs to be vastly updated. <declaration>, which is
parsed with the CSS Syntax algorithm "parse a declaration", now
represents a valid declaration in the context (which is always a style
rule, but this could be configured), which is probably what is expected
but currently unspecified (w3c/csswg-drafts#8795).

Related changes and refactoring will come in future commits.