Properly handle the media provided to create a CSSStyleSheet
CSSOM is assumed to be wrong, since it wants to serialize a media query list
from MediaList, which would represent "the attribute".

When creating a constructed style sheet, the provided MediaList is now reflected
bidirectionally.

When creating a non-constructed style sheet, we expect the owner node to set
CSSStyleSheet.media.mediaText whenever its media attribute is updated.

w3c/csswg-drafts#11778