Remove overrides of general functional notations
The grammar of their value was ignored. For example, empty :is()/:where() was
valid, which may be unexpected. w3c/csswg-drafts#11808

Fixing this problem reaveled that defining <any-value> as a strict or non-strict
comma-containing production prevents specifying {}-block and commas in some
conditions. These rules are ignored for them. w3c/csswg-drafts#11042