Define CSSFunctionDeclarations and CSSPageDeclarations
Also fixes CSSNestedDeclarations and CSSFontFeatureValuesRule (follow-up of
fc5f64c).

CSS*Declarations represent declarations appearing after a rule and expose a
style attribute as a subclass of CSSStyleDeclaration, defined with the
descriptors or properties allowed in the parent rule.

CSSPageDeclarations is unspecified so this is a speculation on the outcome of
w3c/csswg-drafts#11272.

CSSFontFeatureValuesRules hoist declarations for font-display, which will
presumably be removed (w3c/csswg-drafts#2973).