Upgrade dependencies
- <attr()>: updated grammar
- <control-value()>: new function
- <random()>: updated grammar
- <random-caching-options>: updated grammar
- slider-orientation: new property

There are also a bunch of new pseudos from CSS Forms.

control-value() should be an arbitrary substitution (w3c/csswg-drafts#11860),
which is not implemented in this commit, and it should take a type as a keyword
(w3c/csswg-drafts#11842).

random() might now be simplified at parse time when <random-caching-options> is
specified as a <number>, which is not implemented in this commit.