Define a rule in context before parsing its prelude
Prior to this commit and 62c87e1, a rule, declaration, function, was defined in
`context` only when parsing its associated value. It is now also defined when
parsing the prelude of a rule and the name of a function.

The intention was to consider them as execution contexts (scopes) rather than
grammars (nodes) but in other programming languages, the arguments and value of
a function are actually defined in the same scope.

Comparing CSS to other programming languages may not hold any further. In CSS,
most expressions are defined by the language rather than the user. The name of
the structure defines its grammar but it is also part of the production, which
implicitly accepts aliases and mappings.