Match predefined function names case-insensitively
This commit fixes some case-sensitives operations on predefined function names,
which are no longer normalized to lowercase during tokenization since 62c87e1.

For performance reasons, substitutions are not parsed like other functions so
there are now tests in uppercase for all of them.

Some checks are not implemented ideally and will be revisited later.