Do not try consuming a consumed declaration value as a qualified rule
It was already applied for custom properties, which can never make a valid rule.
Since a positioned block prevents consuming a declaration value for a standard
property, the only inputs that would potentially make a valid rule would be like
`name: {...}`, which is very unlikely to become valid anytime soon.