Minor refactoring
The refactorized implementation was brittle because it depended on the relative
color syntax always being the top-level value definition of color functions,
which might not always be true.

The new implementation may slightly increase time complexity but significantly
decrease code complexity because it is the most idiomatic way to check that a
source color has been specified.