Optimize (and fix) parsing <*-in-parens> sub-productions
Parse errors are now returned rather than parse failures. These parse errors are
replaced with the result from parsing against <general-enclosed>, now regardless
of the nesting level of the condition to prevent nested condition combinations
from evaluating to false, like "(not (unknown))".

This is an oversight in 73597c5.

This commit also makes parseDeclaration() return null rather than Error when the
declaration is invalid regardless of its value (if any), to allow matching
<declaration> | <ident>" with "inline-size".