Allow replacing <dimension> with a numeric substitution
Before this commit, math functions could only replace CSS dimension types like
<length>, <angle>, etc... which are all defined as <dimension>. This type was
directly used for type checking the math function.

But they must also be allowed to replace a standalone <dimension> in <mf-value>
and since f2f2a0f, in <progress-source> and <input-position>.

<mf-value> did not need type checking because it is entirely validated against
the grammar of the corresponding <mf-name> in a postprocessing action.

<progress-source> and <input-position> have <dimension> replaced with <calc-sum>
type checked in a postprocessing action. This will be removed in next commit.

A math function replacing a standalone <dimension> must simply match one the CSS
dimension types that a math function can resolve to.