Extract parsing of declaration value substitutions
Ie. an arbitrary substitution containing value and a whole value substitution
(including a CSS-wide keyword).

This is a preamble to an ongoing refactoring of the context model that will
prevent a declaration and its value from having exactly the same definition, but
avoids introducing "declaration-value" as a new definition type just for this.

But this may be reverted later because "parseCSSGrammar()" is no longer able to
return a valid result when receiving one of these substitutions and the grammar
of a declaration value, as required by "CSS.supports(property, value)".