Refactorize context
The previous context model exposed the closest function and block, but did not
allow determining which was the closest. Representing a block is useless at the
moment because it has no semantic value, but it defines the token marking the
end of its value, which will be required in a future commit.

Overall, the new model is simpler, more functional and idiomatic.