Prioritize CSS-wide keywords over declaration value definition
This is not specified very explicitly but "CSS-wide keywords can be used in
custom properties, with the same meaning as in any another property"
(CSS Variables), which requires them to have priority over their value
definition, in order to be computed before substitution.

CSS Values 5 confirms that "var() function can’t get a CSS-wide keyword from the
custom property itself — ​if you tried to specify that, like `--foo: initial;`,
it would just trigger explicit defaulting for the custom property".