Refactorize substitution functions parsing
All substitution functions are now parsed more consistently.

The name is parsed against a grammar instead of simply matching an identifier,
which adds support for any aliased/mapped names.

The context is then validated to return an error when it is missing a dependency
for the substitution, except for arbitrary substitutions, to allow declaration
value definitions like @property/initial-value to match them.

The input is then parsed against the corresponding production, to allow applying
specific transition actions.