Skip parsing whole value substitution in invalid contexts
Since 824119b, parsing substitution functions return an error after matching the
function name when the context is invalid, except for arbitrary substitutions,
which must be accepted in @property/initial-value. But whole value substitutions
must also be accepted in this context.

So this commit makes the exception a rule, excluding for numeric substitutions.