Parse the whole input as a list of tokens
This allows consuming component values and validating the grammar in a single
pass, which allows interpreting bad tokens as parse errors even when nested,
and parsing any type of value with parseGrammar().

This deviates from CSS Syntax procedures that consume a prelude, a declaration
value, a standalone value, as a list of component values, but they are written
to prioritize understandability over efficiency. There should be no observable
difference in the output.

Since the related changes do not allow for easy diffing, this commit also comes
with a few unrelated but non-important fixes and refactorings.