::details-content and ::part() are element-backed
Before this commit, ::details-content was not defined as element-backed, and all
pseudos could be compounded to ::part(). Both are now defined as element-backed
and can be compounded with pseudos that do not expose structural informations.

w3c/csswg-drafts#10786 resolved on making these pseudos invalid at parse time
but this is still unspecified.