Pass restrictions to arguments of any functional pseudo-class
The specification wants all logical pseudo-classes to pass restrictions at their
position to their arguments.

This commit extends this requirement to any functional pseudo-class.

In fact, this already applied when they were compounded to a pseudo-element and
they took a complex selector or a pseudo-class that cannot be compounded to this
pseudo-element.

w3c/csswg-drafts#12718